pink bandana party
changed color of "course-tabs a" to howest pink because based god (it's
realy about making it a little bit easier to read). + changed #4bc9f2
in the variables.less to @color-howest this way it's easier to see
which colors those are. Now primary and secondary colors should be a
bit different but this has not broken anything so yeah #yolo.